refactor: Change Vector and XYPosition to XY
