+++ logic/Preparation/Interface/IGameObj.cs
@@ -6,8 +6,8 @@ namespace Preparation.Interface
     {
         public GameObjType Type { get; }
         public long ID { get; }
-        public XYPosition Position { get; }  // if Square, Pos equals the center
-        public double FacingDirection { get; }
+        public XY Position { get; }  // if Square, Pos equals the center
+        public XY FacingDirection { get; }
         public bool IsRigid { get; }
         public ShapeType Shape { get; }
         public bool CanMove { get; set; }
