+++ logic/Preparation/Interface/IMoveable.cs
@@ -7,9 +7,9 @@ namespace Preparation.Interface
     {
         object MoveLock { get; }
         public int MoveSpeed { get; }
-        public long Move(Vector moveVec);
+        public long Move(XY moveVec);
         protected bool IgnoreCollide(IGameObj targetObj);                     // 忽略碰撞，在具体类中实现
-        public bool WillCollideWith(IGameObj? targetObj, XYPosition nextPos)  // 检查下一位置是否会和目标物碰撞
+        public bool WillCollideWith(IGameObj? targetObj, XY nextPos)  // 检查下一位置是否会和目标物碰撞
         {
             if (targetObj == null)
                 return false;
@@ -20,7 +20,7 @@ namespace Preparation.Interface
                 return false;
             if (targetObj.Shape == ShapeType.Circle)
             {
-                return XYPosition.Distance(nextPos, targetObj.Position) < targetObj.Radius + Radius;
+                return XY.Distance(nextPos, targetObj.Position) < targetObj.Radius + Radius;
             }
             else  // Square
             {
